requirements.txt
# Setting up a Python virtual environment
# python3 -m venv venv 
# .\venv\Scripts\activate
# $env:FLASK_DEBUG=1


# pip install -r requirements.txt
Flask
flask-restful
flask-jwt-extended
pymongo
python-dotenv

